// src/types/openapi.d.ts
export interface components {
    schemas: {
// ...
        Row: {
            /**
             * @description ID of the row.
             * @example i-tuVwxYz
             */
            id: string;
            /**
             * @description The type of this resource.
             * @enum {string}
             */
            type: "row";
            /**
             * Format: url
             * @description API link to the row.
             * @example https://coda.io/apis/v1/docs/AbCDeFGH/tables/grid-pqRst-U/rows/i-RstUv-W
             */
            href: string;
            /**
             * @description The display name of the row, based on its identifying column.
             * @example Apple
             */
            name: string;
            /**
             * @description Index of the row within the table.
             * @example 7
             */
            index: number;
            /**
             * Format: url
             * @description Browser-friendly link to the row.
             * @example https://coda.io/d/_dAbCDeFGH#Teams-and-Tasks_tpqRst-U/_rui-tuVwxYz
             */
            browserLink: string;
            /**
             * Format: date-time
             * @description Timestamp for when the row was created.
             * @example 2018-04-11T00:18:57.946Z
             */
            createdAt: string;
            /**
             * Format: date-time
             * @description Timestamp for when the row was last modified.
             * @example 2018-04-11T00:18:57.946Z
             */
            updatedAt: string;
            /**
             * @description Values for a specific row, represented as a hash of column IDs (or names with `useColumnNames`) to values.
             * @example {
             *       "c-tuVwxYz": "Apple",
             *       "c-bCdeFgh": [
             *         "$12.34",
             *         "$56.78"
             *       ]
             *     }
             */
            values: {
                [key: string]: components["schemas"]["CellValue"];
            };
        };
// ...
        RowList: {
            items: components["schemas"]["Row"][];
            /**
             * Format: url
             * @description API link to these results
             * @example https://coda.io/apis/v1/docs/AbCDeFGH/tables/grid-pqRst-U/rows?limit=20
             */
            href?: string;
            nextPageToken?: components["schemas"]["nextPageToken"];
            nextPageLink?: components["schemas"]["nextPageLink"] & string;
            nextSyncToken?: components["schemas"]["nextSyncToken"];
        };
// ...
        RowEdit: {
            cells: components["schemas"]["CellEdit"][];
        };
        /** @description Payload for updating a row in a table. */
        RowUpdate: {
            row: components["schemas"]["RowEdit"];
        };
        /** @description The result of a row update. */
        RowUpdateResult: components["schemas"]["DocumentMutateResponse"] & {
            /**
             * @description ID of the updated row.
             * @example i-tuVwxYz
             */
            id: string;
        };
        /** @description Payload for deleting rows from a table. */
        RowsDelete: {
            /**
             * @description Row IDs to delete.
             * @example [
             *       "i-bCdeFgh",
             *       "i-CdEfgHi"
             *     ]
             */
            rowIds: string[];
        };
// ...
        RowsUpsert: {
            rows: components["schemas"]["RowEdit"][];
            /**
             * @description Optional column IDs, URLs, or names (fragile and discouraged), specifying columns to be used as upsert keys.
             * @example [
             *       "c-bCdeFgh"
             *     ]
             */
            keyColumns?: string[];
        };
// ...
    };
// ...
}